Initialize resolution project codebase
The initial experiment for remote resolution was opened as a pull
request against the experimental repo.

This commit brings across some of the core bits of machinery from that
experiment to form the basis of resolution. A very basic git resolver is
included that can request files from public git repositories.