Add initial ci pipelines
Prior to this commit we didn't have any ci pipelines defined in the
repo.

This just adds an initial basic pipeline that clones, builds and unit
tests the tektoncd/remote repo. These aren't in active use in our ci
right now but is a foothold on the way to something better defined.