Document version of pipelines required more specifically
At time of writing there's an in-progress pull request for pipelines
that includes the integration with this project. This isn't
documented in the resolution repo but it probably should be.

This commit updates the requirements sections of the project's READMEs
to indicate the exact version of pipelines that needs to be deployed.