Add a Developer Howto for writing a Resolver
Prior to this commit the specifics of writing a Resolver were difficult
to figure out when working from scratch.

This commit adds a developer HowTo doc which walks a new dev through the
process of writing a Resolver. The result of this HowTo is
also added as a template. A developer can copy this to form the basis
of their own Resolver.