Tidy up doc licenses and add howto to docs index
Prior to this commit we didn't link to the new howto guide for
developers from the `docs/` README.md file.

This commit adds a link to the howto from `docs/README.md` and adds a
line separator to all docs so that their license is clearly demarcated
from their core content.