Add list of resolvers to main README
Prior to this commit the list of resolvers that a user could install was
not easy to find.

This commit updates the root README.md with a list of the resolvers that
can be installed right now. It also points users who want to develop a
resolver to the how-to-write-a-resolver.md doc.

Also updated: parameters for each resolver are moved to the top of their
respective READMEs for quicker reference.