Remove namespace param from bundle resolver
Prior to this commit ResolutionRequests for bundle entries relied on the
namespace of the originating TaskRun/PipelineRun to be given via a
param.

After this commit the namespace of the resolution request is available
to all resolvers using the `common.RequestNamespace(ctx)` helper. This
removes the need to pass a namespace explicitly as a parameter when
requesting a bundle. A service account name is still required in order
to construct a k8schain instance.