Remove reference to in-progres pipelines PR
Prior to this commit installation instructions sent users to an
in-progress pull request in the pipelines repo. Now that initial support
for resolution is merged for pipelineRefs that PR can be replaced with
installing pipelines directly from its main branch.

This commit updates installation instructions and our e2e test script to
pull pipelines in from its main branch.