E2E test Tekton Pipelines + Bundle Resolver
Prior to this commit the e2e tests in this repo only checked the
behaviour of ResolutionRequests. This meant that we werent testing
scenarios involving Tekton Pipelines, only the contract that Pipelines
adheres to.

This commit introduces a bundles E2E test that deploys a registry, publishes
a bundle to it, and executes a PipelineRun using the bundles resolver.
This test is a very close cousin to an equivalent test in Pipelines that
uses Pipeline's built-in bundle resolution logic:
https://github.com/tektoncd/pipeline/blob/main/test/tektonbundles_test.go

In addition to the new test code a bug fix has been made to the
framework.Resolver reconciler that prevented timed out requests from
being updated.