Deploy a test registry and bundle before e2e tests
Prior to this commit the pipeline_bundles_test deployed its own test
registry and a bundle. This took a long time and really slowed down the
test.

This commit attempts to create the registry and test bundle before
executing tests, hopefully speeding things up and allowing all tests to
share the same bundleRef.