Rename "myresolver" to "demoresolver" in docs
Prior to this commit the docs provided a working resolver template for
developers that was named "myresolver".

This commit renames the default template from "myresolver" to
"demoresolver" and tweaks a bit of the template's code to make the
`GetName` and `GetSelector` methods a bit more idiomatic, dropping the
"resolver" portion of the name and they type.