Support resolver configmaps and custom timeouts
Prior to this commit all resolvers ran with a 30 second timeout for
resolution requests and didn't have a way to watch a configmap for
admin-level configuration.

This commit does several things:

Sets the default timeout to 1 minute to match the global timeout
enforced by the "core" ResolutionRequest reconciler.
Adds an optional interface for resolvers to implement that lets them
set their own timeouts per-request.
To support configurable timeouts resolvers can now optionally name a
a configmap to watch for admin settings. This is implemented as an
optional interface that a resolver can implement.
The git resolver is updated to add support for a settings configmap
that will eventually contain lots of interesting git-related settings but
today is solely used to allow admins to set a custom timeout for git
fetches.