User Provided defaults for git, bundle and hub resolvers
Prior to this, users had to explicitly provide all the parameters required for the remote resolvers. Now, we can add default parameter values during installation and not have to provide them every single time, thus helping reduce verbosity. The defaulted params for git are `url` and `branch` while those for bundle are `service account` and `kind`.